requirements.txt
alembic==1.1.0
aniso8601==8.0.0
attrs==19.1.0
bcrypt==3.1.4
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Migrate==2.2.1
flask-restplus==0.10.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==3.0.2
Mako==1.1.0
MarkupSafe==1.1.1
peewee==3.10.0
pycparser==2.19
pycrypto==2.6.1
PyJWT==1.6.4
pyrsistent==0.15.4
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.2
requests==2.19.1
six==1.12.0
SQLAlchemy==1.3.8
urllib3==1.23
Werkzeug==0.15.6
